update to support gridfitgpu_par
when using a GPU model and more than 1 GPU is available, now uses a parallelized version of gridfitgpu to simultaneously estimate models across GPUs. At present, does not provide a n_GPU-fold improvement in performance, mostly due to inefficiencies in slice-based processing that will be fixed in a future update.